Rename stuff from virtuoso
